Enable geo check build if vecgeom is disabled

+++ app/geo-check/GCheckKernel.cc
@@ -71,7 +71,7 @@ GCheckOutput run_cpu(const SPConstGeo&          params,
                     step.distance,
                     geo.find_next_step().distance,
                     physid(geo),
-                    geo.volume_id().get());
+                    static_cast<int>(geo.volume_id().get()));
             }
         }
         else
